Inventory Upgrades and ScreenItems
Tied the InventoryItem used by the backpack to the screen via the ScreenItem. Intended to be listed per room the player searches. I stubbed out a "display_item" function that I'll flesh out later so conditions must be met  before the item is shown in the room.
Lastly I've attempted to set up items returning to their position when they aren't dropped on the backpack, but I haven't gotten it working yet.