Scene Interactables
Minor linting in script.rpy.
Started the data classes for clickable scene elements like doors, characters, and in a future pass the ScreenItem.
Locations can have a list of required flags to lock them behind plot points or reveal them only if the player has an item. They'll primarily be used as doors to transition to other scenes.